fraud-detection/requirements.txt
# Core ML
xgboost==2.0.3
scikit-learn==1.4.2
imbalanced-learn==0.12.3
shap==0.45.1
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Dev / CI
python-dotenv==1.0.1
